Initial SDK checkin
Initial checkin of SDK files (all bunched together since arcanist wasn't
making me a diff).
Contains model classes for posting data and endpoints for creating and
retrieving cases. Still pending is the endpoint for creating webhooks,
and handlers for webhooks (validate auth, parse format). Also decided to
move up to PHP 5.3 (which is old enough that it is no longer getting
security updates, but new enough there are still a few things that can't
use it), and am in the process of adding namespaces. Also adding
validation code to verify types, formats and ranges of post data to help
customers diagnose issues when the Signifyd API returns a malformed post
error.